+++ produit.hpp
@@ -14,7 +14,7 @@ struct Produit
     // Constructeur par defaut
     Produit();
     // constructeur
-    Produit(int _idProduit, float _prix, int _idEntreprise, int _quantite, int _qualite);
+    Produit(int _idProduit, float _prix, int _idEntreprise, int _quantite, float _qualite);
 };
 
 #endif
